2022.05.30

+++ models/transformer/perceiver.py
@@ -21,8 +21,9 @@ def default(val, d):
 
 def cache_fn(f):
     cache = dict()
+
     @wraps(f)
-    def cached_fn(*args, _cache = True, key = None, **kwargs):
+    def cached_fn(*args, _cache=True, key=None, **kwargs):
         if not _cache:
             return f(*args, **kwargs)
         nonlocal cache
@@ -31,25 +32,27 @@ def cache_fn(f):
         result = f(*args, **kwargs)
         cache[key] = result
         return result
+
     return cached_fn
 
 
-def fourier_encode(x, max_freq, num_bands = 4):
+def fourier_encode(x, max_freq, num_bands=4):
     x = x.unsqueeze(-1)
     device, dtype, orig_x = x.device, x.dtype, x
 
-    scales = torch.linspace(1., max_freq / 2, num_bands, device = device, dtype = dtype)
+    scales = torch.linspace(1., max_freq / 2, num_bands, device=device, dtype=dtype)
     scales = scales[(*((None,) * (len(x.shape) - 1)), Ellipsis)]
 
     x = x * scales * pi
-    x = torch.cat([x.sin(), x.cos()], dim = -1)
-    x = torch.cat((x, orig_x), dim = -1)
+    x = torch.cat([x.sin(), x.cos()], dim=-1)
+    x = torch.cat((x, orig_x), dim=-1)
     return x
+
 
 # helper classes
 
 class PreNorm(nn.Module):
-    def __init__(self, dim, fn, context_dim = None):
+    def __init__(self, dim, fn, context_dim=None):
         super().__init__()
         self.fn = fn
         self.norm = nn.LayerNorm(dim)
@@ -61,19 +64,19 @@ class PreNorm(nn.Module):
         if exists(self.norm_context):
             context = kwargs['context']
             normed_context = self.norm_context(context)
-            kwargs.update(context = normed_context)
+            kwargs.update(context=normed_context)
 
         return self.fn(x, **kwargs)
 
 
 class GEGLU(nn.Module):
     def forward(self, x):
-        x, gates = x.chunk(2, dim = -1)
+        x, gates = x.chunk(2, dim=-1)
         return x * F.gelu(gates)
 
 
 class FeedForward(nn.Module):
-    def __init__(self, dim, mult = 4, dropout = 0.):
+    def __init__(self, dim, mult=4, dropout=0.):
         super().__init__()
         self.net = nn.Sequential(
             nn.Linear(dim, dim * mult * 2),
@@ -87,7 +90,7 @@ class FeedForward(nn.Module):
 
 
 class Attention(nn.Module):
-    def __init__(self, query_dim, context_dim = None, heads = 8, dim_head = 64, dropout = 0.):
+    def __init__(self, query_dim, context_dim=None, heads=8, dim_head=64, dropout=0.):
         super().__init__()
         inner_dim = dim_head * heads
         context_dim = default(context_dim, query_dim)
@@ -95,63 +98,64 @@ class Attention(nn.Module):
         self.scale = dim_head ** -0.5
         self.heads = heads
 
-        self.to_q = nn.Linear(query_dim, inner_dim, bias = False)
-        self.to_kv = nn.Linear(context_dim, inner_dim * 2, bias = False)
+        self.to_q = nn.Linear(query_dim, inner_dim, bias=False)
+        self.to_kv = nn.Linear(context_dim, inner_dim * 2, bias=False)
 
         self.dropout = nn.Dropout(dropout)
         self.to_out = nn.Linear(inner_dim, query_dim)
 
-    def forward(self, x, context = None, mask = None):
+    def forward(self, x, context=None, mask=None):
         h = self.heads
 
         q = self.to_q(x)
         context = default(context, x)
-        k, v = self.to_kv(context).chunk(2, dim = -1)
+        k, v = self.to_kv(context).chunk(2, dim=-1)
 
-        q, k, v = map(lambda t: rearrange(t, 'b n (h d) -> (b h) n d', h = h), (q, k, v))
+        q, k, v = map(lambda t: rearrange(t, 'b n (h d) -> (b h) n d', h=h), (q, k, v))
 
         sim = einsum('b i d, b j d -> b i j', q, k) * self.scale
 
         if exists(mask):
             mask = rearrange(mask, 'b ... -> b (...)')
             max_neg_value = -torch.finfo(sim.dtype).max
-            mask = repeat(mask, 'b j -> (b h) () j', h = h)
+            mask = repeat(mask, 'b j -> (b h) () j', h=h)
             sim.masked_fill_(~mask, max_neg_value)
 
         # attention, what we cannot get enough of
-        attn = sim.softmax(dim = -1)
+        attn = sim.softmax(dim=-1)
         attn = self.dropout(attn)
 
         out = einsum('b i j, b j d -> b i d', attn, v)
-        out = rearrange(out, '(b h) n d -> b n (h d)', h = h)
+        out = rearrange(out, '(b h) n d -> b n (h d)', h=h)
         return self.to_out(out)
+
 
 # main class
 
 class Perceiver(nn.Module):
     def __init__(
-        self,
-        *,
-        num_freq_bands=6,
-        depth=6,
-        max_freq=224.,
-        input_channels = 3,
-        input_axis = 2,
+            self,
+            *,
+            num_freq_bands=6,
+            depth=6,
+            max_freq=224.,
+            input_channels=3,
+            input_axis=2,
 
-        num_latents = 256,
-        latent_dim = 512,
-        cross_heads = 1,
-        latent_heads = 8,
-        cross_dim_head = 64,
-        latent_dim_head = 64,
+            num_latents=256,
+            latent_dim=512,
+            cross_heads=1,
+            latent_heads=8,
+            cross_dim_head=64,
+            latent_dim_head=64,
 
-        num_classes = 1000,
-        attn_dropout = 0.,
-        ff_dropout = 0.,
-        weight_tie_layers = False,
-        fourier_encode_data = True,
-        self_per_cross_attn = 4,
-        final_classifier_head = True,
+            num_classes=1000,
+            attn_dropout=0.,
+            ff_dropout=0.,
+            weight_tie_layers=False,
+            fourier_encode_data=True,
+            self_per_cross_attn=4,
+            final_classifier_head=True,
     ):
         """The shape of the final attention mechanism will be:
         depth * (cross attention -> self_per_cross_attn * self attention)
@@ -186,17 +190,22 @@ class Perceiver(nn.Module):
         self.num_freq_bands = num_freq_bands
 
         self.fourier_encode_data = fourier_encode_data
-        fourier_channels = (input_axis * ((num_freq_bands * 2) + 1)) if fourier_encode_data else 0      # 26
-        input_dim = fourier_channels + input_channels       # 29
+        fourier_channels = (input_axis * ((num_freq_bands * 2) + 1)) if fourier_encode_data else 0  # 26
+        input_dim = fourier_channels + input_channels  # 29
 
         self.latents = nn.Parameter(torch.randn(num_latents, latent_dim))
 
-        get_cross_attn = lambda: PreNorm(latent_dim, Attention(latent_dim, input_dim, heads = cross_heads, dim_head = cross_dim_head, dropout = attn_dropout), context_dim = input_dim)
-        get_cross_ff = lambda: PreNorm(latent_dim, FeedForward(latent_dim, dropout = ff_dropout))
-        get_latent_attn = lambda: PreNorm(latent_dim, Attention(latent_dim, heads = latent_heads, dim_head = latent_dim_head, dropout = attn_dropout))
-        get_latent_ff = lambda: PreNorm(latent_dim, FeedForward(latent_dim, dropout = ff_dropout))
+        get_cross_attn = lambda: PreNorm(latent_dim,
+                                         Attention(latent_dim, input_dim, heads=cross_heads, dim_head=cross_dim_head,
+                                                   dropout=attn_dropout), context_dim=input_dim)
+        get_cross_ff = lambda: PreNorm(latent_dim, FeedForward(latent_dim, dropout=ff_dropout))
+        get_latent_attn = lambda: PreNorm(latent_dim,
+                                          Attention(latent_dim, heads=latent_heads, dim_head=latent_dim_head,
+                                                    dropout=attn_dropout))
+        get_latent_ff = lambda: PreNorm(latent_dim, FeedForward(latent_dim, dropout=ff_dropout))
 
-        get_cross_attn, get_cross_ff, get_latent_attn, get_latent_ff = map(cache_fn, (get_cross_attn, get_cross_ff, get_latent_attn, get_latent_ff))
+        get_cross_attn, get_cross_ff, get_latent_attn, get_latent_ff = map(cache_fn, (
+        get_cross_attn, get_cross_ff, get_latent_attn, get_latent_ff))
 
         self.layers = nn.ModuleList([])
         for i in range(depth):
@@ -207,8 +216,8 @@ class Perceiver(nn.Module):
 
             for block_ind in range(self_per_cross_attn):
                 self_attns.append(nn.ModuleList([
-                    get_latent_attn(**cache_args, key = block_ind),
-                    get_latent_ff(**cache_args, key = block_ind)
+                    get_latent_attn(**cache_args, key=block_ind),
+                    get_latent_ff(**cache_args, key=block_ind)
                 ]))
 
             self.layers.append(nn.ModuleList([
@@ -224,38 +233,37 @@ class Perceiver(nn.Module):
         ) if final_classifier_head else nn.Identity()
 
     def forward(
-        self,
-        data,
-        mask = None,
-        return_embeddings = False
+            self,
+            data,
+            mask=None,
+            return_embeddings=False
     ):
         if self.input_axis == 2:
             data = rearrange(data, 'b c h w -> b h w c')
         b, *axis, _, device, dtype = *data.shape, data.device, data.dtype
         assert len(axis) == self.input_axis, 'input data must have the right number of axis'
 
-
         if self.fourier_encode_data:
             # calculate fourier encoded positions in the range of [-1, 1], for all axis
 
             axis_pos = list(map(lambda size: torch.linspace(-1., 1., steps=size, device=device, dtype=dtype), axis))
-            pos = torch.stack(torch.meshgrid(*axis_pos, indexing = 'ij'), dim = -1)
+            pos = torch.stack(torch.meshgrid(*axis_pos, indexing='ij'), dim=-1)
             enc_pos = fourier_encode(pos, self.max_freq, self.num_freq_bands)
             enc_pos = rearrange(enc_pos, '... n d -> ... (n d)')
-            enc_pos = repeat(enc_pos, '... -> b ...', b = b)
+            enc_pos = repeat(enc_pos, '... -> b ...', b=b)
 
-            data = torch.cat((data, enc_pos), dim = -1)
+            data = torch.cat((data, enc_pos), dim=-1)
 
         # concat to channels of data and flatten axis
 
         data = rearrange(data, 'b ... d -> b (...) d')
 
-        x = repeat(self.latents, 'n d -> b n d', b = b)
+        x = repeat(self.latents, 'n d -> b n d', b=b)
 
         # layers
 
         for cross_attn, cross_ff, self_attns in self.layers:
-            x = cross_attn(x, context = data, mask = mask) + x
+            x = cross_attn(x, context=data, mask=mask) + x
             x = cross_ff(x) + x
 
             for self_attn, self_ff in self_attns:
@@ -274,26 +282,26 @@ class Perceiver(nn.Module):
 
 class PerceiverNotShared(nn.Module):
     def __init__(
-        self,
-        *,
-        num_freq_bands=32,
-        depth=6,
-        max_freq=10.,
-        input_channels = 3,
-        input_axis = 2,
-        num_latents = 128,
-        latent_dim = 256,
-        cross_heads = 1,
-        latent_heads = 8,
-        cross_dim_head = 16,
-        latent_dim_head = 16,
-        num_classes = 1000,
-        attn_dropout = 0.,
-        ff_dropout = 0.,
-        weight_tie_layers = False,
-        fourier_encode_data = True,
-        self_per_cross_attn = 1,
-        final_classifier_head = True
+            self,
+            *,
+            num_freq_bands=32,
+            depth=6,
+            max_freq=10.,
+            input_channels=3,
+            input_axis=2,
+            num_latents=128,
+            latent_dim=256,
+            cross_heads=1,
+            latent_heads=8,
+            cross_dim_head=16,
+            latent_dim_head=16,
+            num_classes=1000,
+            attn_dropout=0.,
+            ff_dropout=0.,
+            weight_tie_layers=False,
+            fourier_encode_data=True,
+            self_per_cross_attn=1,
+            final_classifier_head=True
     ):
         """The shape of the final attention mechanism will be:
         depth * (cross attention -> self_per_cross_attn * self attention)
@@ -328,17 +336,22 @@ class PerceiverNotShared(nn.Module):
         self.num_freq_bands = num_freq_bands
 
         self.fourier_encode_data = fourier_encode_data
-        fourier_channels = (input_axis * ((num_freq_bands * 2) + 1)) if fourier_encode_data else 0      # 26
-        input_dim = fourier_channels + input_channels       # 29
+        fourier_channels = (input_axis * ((num_freq_bands * 2) + 1)) if fourier_encode_data else 0  # 26
+        input_dim = fourier_channels + input_channels  # 29
 
         self.latents = nn.Parameter(torch.randn(num_latents, latent_dim))
 
-        get_cross_attn = lambda: PreNorm(latent_dim, Attention(latent_dim, input_dim, heads = cross_heads, dim_head = cross_dim_head, dropout = attn_dropout), context_dim = input_dim)
-        get_cross_ff = lambda: PreNorm(latent_dim, FeedForward(latent_dim, dropout = ff_dropout))
-        get_latent_attn = lambda: PreNorm(latent_dim, Attention(latent_dim, heads = latent_heads, dim_head = latent_dim_head, dropout = attn_dropout))
-        get_latent_ff = lambda: PreNorm(latent_dim, FeedForward(latent_dim, dropout = ff_dropout))
+        get_cross_attn = lambda: PreNorm(latent_dim,
+                                         Attention(latent_dim, input_dim, heads=cross_heads, dim_head=cross_dim_head,
+                                                   dropout=attn_dropout), context_dim=input_dim)
+        get_cross_ff = lambda: PreNorm(latent_dim, FeedForward(latent_dim, dropout=ff_dropout))
+        get_latent_attn = lambda: PreNorm(latent_dim,
+                                          Attention(latent_dim, heads=latent_heads, dim_head=latent_dim_head,
+                                                    dropout=attn_dropout))
+        get_latent_ff = lambda: PreNorm(latent_dim, FeedForward(latent_dim, dropout=ff_dropout))
 
-        get_cross_attn, get_cross_ff, get_latent_attn, get_latent_ff = map(cache_fn, (get_cross_attn, get_cross_ff, get_latent_attn, get_latent_ff))
+        get_cross_attn, get_cross_ff, get_latent_attn, get_latent_ff = map(cache_fn, (
+        get_cross_attn, get_cross_ff, get_latent_attn, get_latent_ff))
 
         self.layers = nn.ModuleList([])
         for i in range(depth):
@@ -349,8 +362,8 @@ class PerceiverNotShared(nn.Module):
 
             for block_ind in range(self_per_cross_attn):
                 self_attns.append(nn.ModuleList([
-                    get_latent_attn(**cache_args, key = block_ind),
-                    get_latent_ff(**cache_args, key = block_ind)
+                    get_latent_attn(**cache_args, key=block_ind),
+                    get_latent_ff(**cache_args, key=block_ind)
                 ]))
 
             self.layers.append(nn.ModuleList([
@@ -366,38 +379,37 @@ class PerceiverNotShared(nn.Module):
         ) if final_classifier_head else nn.Identity()
 
     def forward(
-        self,
-        data,
-        mask = None,
-        return_embeddings = False
+            self,
+            data,
+            mask=None,
+            return_embeddings=False
     ):
         if self.input_axis == 2:
             data = rearrange(data, 'b c h w -> b h w c')
         b, *axis, _, device, dtype = *data.shape, data.device, data.dtype
         assert len(axis) == self.input_axis, 'input data must have the right number of axis'
 
-
         if self.fourier_encode_data:
             # calculate fourier encoded positions in the range of [-1, 1], for all axis
 
             axis_pos = list(map(lambda size: torch.linspace(-1., 1., steps=size, device=device, dtype=dtype), axis))
-            pos = torch.stack(torch.meshgrid(*axis_pos, indexing = 'ij'), dim = -1)
+            pos = torch.stack(torch.meshgrid(*axis_pos, indexing='ij'), dim=-1)
             enc_pos = fourier_encode(pos, self.max_freq, self.num_freq_bands)
             enc_pos = rearrange(enc_pos, '... n d -> ... (n d)')
-            enc_pos = repeat(enc_pos, '... -> b ...', b = b)
+            enc_pos = repeat(enc_pos, '... -> b ...', b=b)
 
-            data = torch.cat((data, enc_pos), dim = -1)
+            data = torch.cat((data, enc_pos), dim=-1)
 
         # concat to channels of data and flatten axis
 
         data = rearrange(data, 'b ... d -> b (...) d')
 
-        x = repeat(self.latents, 'n d -> b n d', b = b)
+        x = repeat(self.latents, 'n d -> b n d', b=b)
 
         # layers
 
         for cross_attn, cross_ff, self_attns in self.layers:
-            x = cross_attn(x, context = data, mask = mask) + x
+            x = cross_attn(x, context=data, mask=mask) + x
             x = cross_ff(x) + x
 
             for self_attn, self_ff in self_attns:
